+++ projects/packages/my-jetpack/_inc/components/connected-product-card/index.jsx
@@ -9,7 +9,7 @@ import ProductCard from '../product-card';
 const ConnectedProductCard = ( { admin, slug } ) => {
 	const { isRegistered, isUserConnected } = useConnection();
 	const { detail, status, activate, deactivate, isFetching } = useProduct( slug );
-	const { name, description, manageUrl } = detail;
+	const { name, description, manageUrl, requiresUserConnection } = detail;
 
 	const navigateToConnectionPage = useMyJetpackNavigate( '/connection' );
 
@@ -25,14 +25,20 @@ const ConnectedProductCard = ( { admin, slug } ) => {
 	/*
 	 * Redirect only if connected
 	 */
-	const callOnlyIfAllowed = callback => () => {
-		if ( ! isRegistered || ! isUserConnected ) {
+	const handleActivate = useCallback( () => {
+		if ( ( ! isRegistered || ! isUserConnected ) && requiresUserConnection ) {
 			navigateToConnectionPage();
 			return;
 		}
 
-		callback();
-	};
+		activate();
+	}, [
+		activate,
+		isRegistered,
+		isUserConnected,
+		requiresUserConnection,
+		navigateToConnectionPage,
+	] );
 
 	const Icon = getIconBySlug( slug );
 
@@ -46,8 +52,8 @@ const ConnectedProductCard = ( { admin, slug } ) => {
 			isFetching={ isFetching }
 			onDeactivate={ deactivate }
 			slug={ slug }
-			onActivate={ activate }
-			onAdd={ callOnlyIfAllowed( navigateToAddProductPage ) }
+			onActivate={ handleActivate }
+			onAdd={ navigateToAddProductPage }
 			onManage={ onManage }
 			onFixConnection={ navigateToConnectionPage }
 		/>