+++ projects/packages/my-jetpack/src/class-initializer.php
@@ -30,10 +30,10 @@ class Initializer {
 	 *
 	 * @var string
 	 */
-	const PACKAGE_VERSION = '2.5.0';
+	const PACKAGE_VERSION = '2.6.0-alpha';
 
 	/**
-	 * Initialize My Jetapack
+	 * Initialize My Jetpack
 	 *
 	 * @return void
 	 */